+++ scripts/data.js
@@ -1,12 +1,19 @@
+const firebase = require("firebase/app");
+require("firebase/database");
+import { displayData } from "./dom";
+
 let clients = [];
 
-const database = firebase.database();
-const clientsRef = database.ref("clients");
+export function getData() {
+  console.log("getData()");
+  const database = firebase.database();
+  const clientsRef = database.ref("clients");
 
-clientsRef.on("value", snapshot => {
-  clients = convertObjToArray(snapshot.val());
-  displayData(clients);
-});
+  clientsRef.on("value", snapshot => {
+    clients = convertObjToArray(snapshot.val());
+    displayData(clients);
+  });
+}
 
 function convertObjToArray(object) {
   return Object.keys(object).map(key => {
@@ -16,3 +23,5 @@ function convertObjToArray(object) {
     };
   });
 }
+
+export { clients };
